requirements.txt
# Core dependencies for Vexa Polish LLM
numpy>=1.21.0
pyyaml>=5.4.1
requests>=2.26.0
tqdm>=4.62.0

# GPU support (CUDA)
pycuda>=2021.1

# Web interface and API
flask>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Evaluation metrics
nltk>=3.8.0
